requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
imageio==2.31.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langid==1.1.6
lazy_loader==0.2
MarkupSafe==2.1.3
networkx==3.1
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
python-dotenv==1.0.0
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
tifffile==2023.4.12
urllib3==2.0.3
Werkzeug==2.3.6
